ngebuat utility untuk data layer sebagai daily alarm setter, selain itu ngubah juga tipe primary key pada daily notification biar jadi integer

+++ app/src/main/java/com/example/gratefulnote/daily_notification/data/repository/DailyNotificationRepository.kt
@@ -25,7 +25,7 @@ class DailyNotificationRepository(
                     hour = hour,
                     minute = minute,
                 )
-            ).single()
+            ).single().toInt()
             emit(ResponseWrapper.ResponseSucceed(id))
         } catch (t : Throwable){
             emit(ResponseWrapper.ResponseError(t))
