+++ app/src/main/java/com/example/gratefulnote/daily_notification/domain/repository/IDailyNotificationRepository.kt
@@ -5,7 +5,7 @@ import com.example.gratefulnote.database.DailyNotificationEntity
 import kotlinx.coroutines.flow.Flow
 
 interface IDailyNotificationRepository {
-    fun createNewDailyNotification(hour : Int, minute : Int) : Flow<ResponseWrapper<Long>>
+    fun createNewDailyNotification(hour : Int, minute : Int) : Flow<ResponseWrapper<Int>>
     fun updateDailyNotification(dailyNotification: DailyNotificationEntity) : Flow<ResponseWrapper<Nothing>>
     fun deleteDailyNotification(vararg dailyNotifications : DailyNotificationEntity) : Flow<ResponseWrapper<Nothing>>
     fun getAllDailyNotification() : Flow<ResponseWrapper<List<DailyNotificationEntity>>>
