+++ app/src/main/java/com/example/gratefulnote/daily_notification/domain/service/IDailyNotificationManager.kt
@@ -5,7 +5,7 @@ import com.example.gratefulnote.database.DailyNotificationEntity
 import kotlinx.coroutines.flow.Flow
 
 interface IDailyNotificationManager {
-    suspend fun addNewDailyNotification(hour : Int, minute : Int) : Flow<ResponseWrapper<Long>>
+    suspend fun addNewDailyNotification(hour : Int, minute : Int) : Flow<ResponseWrapper<Int>>
     suspend fun toogleDailyNotification(dailyNotification: DailyNotificationEntity) : Flow<ResponseWrapper<Nothing>>
     suspend fun deleteDailyNotification(dailyNotifications: List<DailyNotificationEntity>) : Flow<ResponseWrapper<Nothing>>
     fun getAllDailyNotification() : Flow<ResponseWrapper<List<DailyNotificationEntity>>>
